feat: move macos components into separate modules, and improve error handling
