+++ src/error.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum Error {
+    Placeholder
+}
